src/requirements.txt
Flask==2.2.2
nltk==3.7
numpy==1.23.2
pandas==1.4.3
PyYAML==6.0
requests==2.28.1
scikit_learn==1.1.2
spacy==3.4.1
